Fixed bug in LinearEq that somehow got through AutoGrading. Used Integer.parseInt for a double value. Changed it to Double.parseDouble.

+++ LinearEq.java
@@ -6,9 +6,9 @@
  */
 public class LinearEq {
 	public static void main(String[] args) {
-		double a = Integer.parseInt(args[0]);
-		double b = Integer.parseInt(args[1]);
-		double c = Integer.parseInt(args[2]);
+		double a = Double.parseDouble(args[0]);
+		double b = Double.parseDouble(args[1]);
+		double c = Double.parseDouble(args[2]);
 
 		double x = (c - b) / a; // ax + b = c
 		System.out.println(a + " * x + " + b + " = " + c + "\nx = " + x);
